requirements.txt
Django==1.8
PIL==1.1.7
Pillow==2.8.1
django-autoslug-field==0.2.3
django-extensions==1.5.2
django-filter==0.9.2
django-hvad==1.2.0
django-rosetta==0.7.6
djangorestframework==3.1.1
djrill==1.3.0
easy-thumbnails==2.2
microsofttranslator==0.5
polib==1.0.6
psycopg2==2.6
requests==2.6.2
sh==1.11
six==1.9.0
wsgiref==0.1.2
